refactor(auth): Extract login and register pages into reusable UI components
- Replaced inline form logic in LoginPage and CreateUserPage with shared UI components.
- Added success notification after user registration and automatic login + redirect.
- Updated routing to use `ProfileUserPage` instead of deprecated `ProfilePage`.
- Applied conditional home page link rendering based on login state.
- Organized styles and storybook files to match new component structure.

This refactor improves code reuse, simplifies test coverage, and unifies styling across authentication flows.
