+++ packages/quiz/src/pages/CreateUserPage/components/CreateUserPageUI/CreateUserPageUI.module.scss
@@ -1,0 +1,22 @@
+@use '../../../../styles/helpers';
+@use '../../../../styles/variables';
+
+.createUserForm {
+  display: flex;
+  flex-direction: column;
+
+  @include helpers.mobile {
+    row-gap: variables.$mobile-narrow-layout-gutter;
+    width: 100%;
+  }
+
+  @include helpers.tablet {
+    row-gap: variables.$tablet-narrow-layout-gutter;
+    width: variables.$mobile-width;
+  }
+
+  @include helpers.desktop {
+    row-gap: variables.$desktop-narrow-layout-gutter;
+    width: variables.$mobile-width;
+  }
+}
