+++ packages/quiz/src/pages/CreateUserPage/components/index.ts
@@ -1,0 +1,5 @@
+export type {
+  CreateUserPageUIProps,
+  CreateUserFormFields,
+} from './CreateUserPageUI'
+export { default as CreateUserPageUI } from './CreateUserPageUI'
