+++ packages/quiz/src/pages/LoginPage/components/LoginPageUI/index.ts
@@ -1,0 +1,2 @@
+export type { LoginPageUIProps, LoginFormFields } from './LoginPageUI'
+export { default } from './LoginPageUI'
