--- packages/quiz/src/pages/ProfilePage/components/index.ts
@@ -1,5 +1,0 @@
-export type {
-  UpdateUserProfileFormFields,
-  ProfilePageUIProps,
-} from './ProfilePageUI'
-export { default as ProfilePageUI } from './ProfilePageUI'
