--- packages/quiz/src/pages/ProfilePage/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ProfilePage'
