+++ packages/quiz/src/pages/ProfileUserPage/components/ProfileUserPageUI/ProfileUserPageUI.module.scss
