+++ packages/quiz/src/pages/ProfileUserPage/components/ProfileUserPageUI/index.ts
@@ -1,0 +1,5 @@
+export type {
+  UpdateUserProfileFormFields,
+  ProfileUserPageUIProps,
+} from './ProfileUserPageUI.tsx'
+export { default } from './ProfileUserPageUI.tsx'
