+++ packages/quiz/src/pages/ProfileUserPage/components/index.ts
@@ -1,0 +1,5 @@
+export type {
+  UpdateUserProfileFormFields,
+  ProfileUserPageUIProps,
+} from './ProfileUserPageUI'
+export { default as ProfileUserPageUI } from './ProfileUserPageUI'
