+++ packages/quiz/src/pages/ProfileUserPage/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ProfileUserPage'
