+++ packages/quiz/src/pages/index.ts
@@ -7,7 +7,7 @@ export { default as GameResultsPage } from './GameResultsPage'
 export { default as HomePage } from './HomePage'
 export { default as JoinPage } from './JoinPage'
 export { default as LoginPage } from './LoginPage'
-export { default as ProfilePage } from './ProfilePage'
+export { default as ProfileUserPage } from './ProfileUserPage'
 export { default as QuizCreatorPage } from './QuizCreatorPage'
 export { default as QuizDetailsPage } from './QuizDetailsPage'
 export { default as ProfileQuizzesPage } from './ProfileQuizzesPage'
